PPP/pcParts/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
pytz==2024.2
setuptools==75.1.0
sqlparse==0.5.2
wheel==0.44.0
